Enable Continuous Deployment
This commit enables continuous deployment of this project using
Travis CI. Any successful build on the `master` branch will be
deployed. The documentation will be pushed to `gh-pages`, the build
will be published to npm (as a pre-release by default, as a full
release if the git tag matches the package version).