Drop support for `customTypingsDir`
Passing a custom typings dir forced the use of `typeRoots`. Setting this option overrides the default resolution behavior. This default behavior is relatively complex, especially when using Yarn's Plug'n'Play. By dropping support for `customTypingsDir`, we ensure that the default resolution is always applied. The current state of the TypeScript ecosystem makes it also fairly rare to actually need to embed custom typings. A workaround to continue supporting them is to use Yarn's `file:` protocol.